backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.8.1
Werkzeug==2.3.7
python-dotenv==1.0.0
geopy==2.3.0
gunicorn==21.2.0
